Remove Matilda's co-monarchy
The co-monarchy mechanic is broken. The co-monarchy is kept after the ruler's death since the game assumes a co-monarch will be the heir, so no checks were put in place to remove it for other reasons.

Beatrix is still kept on Matilda's court for historicity, but her traits are reverted to vanilla's to make her more useful gameplay-wise for Matilda since she can't be a Chancellor but can be a Spymaster.